@@ -1,15 +1,28 @@
 'use client'
-import React from "react";
-import { useRouter } from 'next/router';
+import React, { useEffect, useState } from "react";
+import { usePathname } from 'next/navigation';
 
 const Domains = () => {
 
-    const router = useRouter();
-    console.log(router);
+    const [domain, setDomain] = useState('');
 
+    const path = usePathname();
+    
+    useEffect(() => {
+        const pathAfterDomains = path.split('/domains/')[1];
+        if (pathAfterDomains) {
+            setDomain(pathAfterDomains);
+        }
+
+        console.log(pathAfterDomains);
+
+        
+    }, [path]);
+    
+    
     return (
         <div>
-            <h1>hiiii</h1>
+            <h1>redirecting to: {domain}</h1>
         </div>
     )
 }
